requirements.txt
matplotlib==3.10.0
numpy==1.26.4
pandas==2.2.3
patsy==1.0.1
pillow==10.4.0
pingouin==0.5.5
psychopy==2024.2.4
scipy==1.15.1
seaborn==0.13.2
statsmodels==0.14.4
